requirements.txt
pytest == 8.3.1
selene >= 2.0.0rc7
allure-pytest == 2.13.5
python-dotenv == 1.0.1
requests == 2.32.3
jsonschema == 4.23.0